fix(orchestrator,cli): break Phase 2 import cycles flagged by check-deps
Phase 2 Task 14 final exit-gate sweep surfaced two circular dependencies
introduced by Tasks 9 (orchestrator OpenAPI registry split) and 11 (CLI
MCP gateway-tools). Both were present from the moment the respective
commits landed but `harness validate` and the per-task gates didn't
catch them — `harness check-deps` runs only at phase exit.

Cycle 1 (orchestrator): registry.ts <-> v1-registry.ts
  - registry.ts had a bottom-of-file `import { buildV1Registry } from
    './v1-registry'` to compose buildV1Document(). v1-registry.ts in
    turn imported buildAuthRegistry from registry.ts.
  - Fix: relocate buildV1Document() into v1-registry.ts (alongside
    buildV1Registry where the dependency belongs). registry.ts now
    has zero imports from v1-registry.ts. generate.ts pivots to
    `from './v1-registry'`.
  - Verified: openapi.yaml is byte-identical pre/post fix.

Cycle 2 (CLI): server.ts <-> tools/gateway-tools.ts
  - server.ts owned the `ToolDefinition` type. tools/gateway-tools.ts
    and tool-tiers.ts pulled the type back via `import type`. Even
    though `import type` erases at runtime, static analysis (and
    future codemods) still walk the edge.
  - Fix: extract `ToolDefinition` to a new ./tool-types.ts module.
    server.ts re-exports for backwards-compat (preserves the public
    `export type { ToolDefinition }` shape that existing tests rely
    on). gateway-tools.ts and tool-tiers.ts now import from
    ./tool-types directly — no edge back to server.ts.

Tests, lint, typecheck, harness validate all still green. harness
check-deps now passes (was 2 issues, now 0).
