+++ packages/orchestrator/src/gateway/openapi/registry.ts
@@ -103,25 +103,9 @@ export function buildAuthDocument(): ReturnType<OpenApiGeneratorV31['generateDoc
   });
 }
 
-// Phase 2 Task 9: composed document covering auth + legacy aliases +
-// bridge primitives. Import is at the bottom to avoid a load-order cycle
-// (v1-registry depends on buildAuthRegistry above).
-//
+// `buildV1Document` lives in `./v1-registry` (alongside `buildV1Registry`)
+// so that registry.ts has zero imports from v1-registry.ts. This avoids the
+// circular dependency that `harness check-deps` flagged at the end of Phase 2.
 // `buildAuthDocument` above is intentionally retained for callers that only
-// want the Phase 1 surface (none in-tree today, but the API contract is
-// public via openapi.yaml generation tooling).
-import { buildV1Registry } from './v1-registry';
-
-export function buildV1Document(): ReturnType<OpenApiGeneratorV31['generateDocument']> {
-  const generator = new OpenApiGeneratorV31(buildV1Registry().definitions);
-  return generator.generateDocument({
-    openapi: '3.1.0',
-    info: {
-      title: 'Harness Gateway API',
-      version: '0.2.0',
-      description:
-        'Hermes Phase 0 — Phase 2: versioned /api/v1/* surface with auth, legacy aliases, and bridge primitives.',
-    },
-    servers: [{ url: 'http://127.0.0.1:8080' }],
-  });
-}
+// want the Phase 1 surface (no in-tree callers today, but the API contract
+// remains importable for downstream auth-only consumers).
